backend/requirements.txt
# Web框架
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# AI & NLP
langchain==0.3.27
langchain-community==0.3.30
langchain-core==0.3.77
chromadb>=0.5.0
openai>=1.0.0

# HTTP请求（调用阿里云百炼API）
httpx>=0.27.0

# 配置管理
python-dotenv>=1.0.0

# 文件处理
python-multipart>=0.0.9
pdfminer.six>=20231228
python-docx>=1.1.0

# 模板引擎
Jinja2>=3.1.0

# 数据处理
pandas>=2.2.0
openpyxl>=3.1.2

# 机器学习
scikit-learn>=1.7.0
numpy>=1.24.0
scipy>=1.15.0

# 类型/数据验证
pydantic>=2.7.0

# 重试机制
tenacity>=8.2.0

# 可选（调试 & 测试）
pytest>=8.2.0
